Refactor revenue tier calculation and update imports
- Moved the `determineRevenueTier` function from `utils/util.js` to `affiliate_helpers.js` for better organization.
- Updated imports across multiple files to reference the new location of `determineRevenueTier`.
- Simplified the `createPrivatePromoCode` function by using shorthand property names for the user parameter.
